Finished fixing the navbar since it was getting stuck when clicking it in small screens.
Added an event listener for the nav element so that when the screen width is smaller then 768 pixels, then it uses an "click" eventListener and when the screen width surpasses the 768, then the screen uses "mauseover" eventListener.

Also made so that every time the screen resizes, an eventListerner will trigger and refresh the page so that the page uses hover and not click.